Fix intermittently failing tests by preventing subscriptions from being mocked
