+++ metaspace/webapp/src/graphqlClient.ts
@@ -1,4 +1,4 @@
-import { ApolloClient, InMemoryCache, defaultDataIdFromObject } from 'apollo-client-preset'
+import { ApolloClient } from 'apollo-client-preset'
 import { BatchHttpLink } from 'apollo-link-batch-http'
 import { WebSocketLink } from 'apollo-link-ws'
 import { setContext } from 'apollo-link-context'
@@ -10,6 +10,7 @@ import config from './config'
 import tokenAutorefresh from './tokenAutorefresh'
 import reportError from './lib/reportError'
 import { get } from 'lodash-es'
+import { makeApolloCache } from './lib/apolloCache'
 
 const graphqlUrl = config.graphqlUrl || `${window.location.origin}/graphql`
 const wsGraphqlUrl = config.wsGraphqlUrl || `${window.location.origin.replace(/^http/, 'ws')}/ws`
@@ -106,9 +107,7 @@ wsClient.use([{
   },
 }])
 
-const wsLink = new WebSocketLink(wsClient);
-(window as any).wsClient = wsClient;
-(window as any).wsLink = wsLink
+const wsLink = new WebSocketLink(wsClient)
 
 const link = errorLink.concat(authLink).split(
   (operation) => {
@@ -120,35 +119,9 @@ const link = errorLink.concat(authLink).split(
   httpLink,
 )
 
-const nonNormalizableTypes: any[] = ['User', 'DatasetUser', 'DatasetGroup', 'DatasetProject']
-
 const apolloClient = new ApolloClient({
   link,
-  cache: new InMemoryCache({
-    cacheRedirects: {
-      Query: {
-        // Allow get-by-id queries to use cached data that originated from other kinds of queries
-        dataset: (_, args, { getCacheKey }) => getCacheKey({ __typename: 'Dataset', id: args.id }),
-        annotation: (_, args, { getCacheKey }) => getCacheKey({ __typename: 'Annotation', id: args.id }),
-        group: (_, args, { getCacheKey }) => getCacheKey({ __typename: 'Group', id: args.groupId }),
-        project: (_, args, { getCacheKey }) => getCacheKey({ __typename: 'Project', id: args.projectId }),
-      },
-    },
-    dataIdFromObject(object: any) {
-      // WORKAROUND: Because of Apollo's aggressive caching, often the current User will be overwritten with results
-      // from other queries. The server side often strips fields based on how they're accessed (the "ScopeRole" logic),
-      // which means these query paths will often return different data with the same IDs:
-      // currentUser -> primaryGroup (always present)
-      // dataset -> submitter -> primaryGroup (null unless admin)
-      // To protect against this, don't allow Users (and possibly other types in the future) to have a dataId,
-      // so that InMemoryCache cannot share data between different queries.
-      if (nonNormalizableTypes.includes(object.__typename)) {
-        return null
-      } else {
-        return defaultDataIdFromObject(object)
-      }
-    },
-  }),
+  cache: makeApolloCache(),
   defaultOptions: {
     query: {
       fetchPolicy: 'network-only',
