+++ metaspace/webapp/src/modules/Datasets/list/DatasetItem.vue
@@ -309,7 +309,6 @@ export default {
         Group: this.dataset.groupApproved ? this.dataset.group : null,
         Projects: this.dataset.projects,
       }
-      console.warn(this.dataset.metadataJson)
       return Object.assign(safeJsonParse(this.dataset.metadataJson), datasetMetadataExternals)
     },
